Extract common blockstats subscription
