+++ src/main.rs
@@ -1,8 +1,10 @@
 mod evm;
+mod stats;
 mod wasm;
 
 // export for use by contract! macro
 use clap::Parser;
+pub use stats::{collect_block_stats, BlockInfo};
 pub use wasm::{InkConstructor, InkMessage};
 
 #[derive(Debug, Parser)]
